+++ CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `memory_bytes` parameter to start session request to allow specifying the memory limit for the session.
+- Added `cpus` parameter to start session request to allow specifying the CPU limit for the session.
+- Added `network_enabled` parameter to start session request to allow specifying whether to enable network for the session.
+- Added integration tests to the project.
+
+### Changed
+
+- Improved patch extraction performance by using shared Docker volumes instead of copying full workspace trees between containers.
+- Changed run container command `sleep` time to 1 hour to avoid containers being removed too early.
+- Changed `persist_workdir` parameter to `ephemeral` in start session request. **Breaking change**
+
+### Removed
+
+- Removed `workdir` parameter from start session request. **Breaking change**
+
 ## [0.3.1] - 2025-12-13
 
 ### Changed